SONARSEC-145: Translate call to operator overload to uCFG method call
