+++ sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/ControlFlowGraph/UcfgBuilder_Instructions.cs
@@ -1639,32 +1639,6 @@ namespace Ns2
         }
 
         [TestMethod]
-        public void Dynamic()
-        {
-            const string code = @"
-namespace Namespace
-{
-    public class Class1
-    {
-        public string Foo(Class1 c)
-        {
-            dynamic dyn = c.Bar;
-                // dyn := __id [ const ]
-
-            if (dyn.User != null)
-            {
-                return ""bar"";
-            }
-
-            return c.ToString();
-                // %0 := object.ToString() [ c ]
-        }
-    }
-}";
-            UcfgVerifier.VerifyInstructions(code, "Foo");
-        }
-
-        [TestMethod]
         public void AddAssignment_LocalVariable()
         {
             const string code = @"
@@ -1930,20 +1904,15 @@ namespace Namespace
         }
 
         [TestMethod]
-        public void NonConcat_BinaryExpressions_AreIgnored()
+        public void AssignmentOperator_UnsupportedTypes()
         {
             const string code = @"
 namespace Namespace
 {
     public class Class1
     {
-        public void Foo(string parameterString)
+        public void Foo()
         {
-            var result = parameterString == null;
-                // result := __id [ const ]
-            result = parameterString != null;
-                // result := __id [ const ]
-
             var i = 42 + 1;
                 // i := __id [ const ]
             i += 1;
@@ -1954,6 +1923,67 @@ namespace Namespace
                 // i := __id [ const ]
             i /= 1;
                 // i := __id [ const ]
+        }
+    }
+}";
+            UcfgVerifier.VerifyInstructions(code, "Foo");
+        }
+
+        [TestMethod]
+        public void Operator_UnsupportedTypes()
+        {
+            const string code = @"
+namespace Namespace
+{
+    public class Class1
+    {
+        public void Foo()
+        {
+            int x;
+            double d;
+            x = 1 + 2;
+                // x := __id [ const ]
+            x = 1 - 2;
+                // x := __id [ const ]
+            x = 1 * 2;
+                // x := __id [ const ]
+            x = 1 / 2;
+                // x := __id [ const ]
+            d = 1.0 / 2;
+                // d := __id [ const ]
+        }
+    }
+}";
+            UcfgVerifier.VerifyInstructions(code, "Foo");
+        }
+
+        [TestMethod]
+        public void Operator_SupportedTypes()
+        {
+            const string code = @"
+namespace Namespace
+{
+    public class Class1
+    {
+        public void Foo(string parameterString, Class1 some, Class1 other)
+        {
+            var result = parameterString == null;
+                // %0 := string.operator ==(string, string) [ string parameterString const ]
+                // result := __id [ %0 ]
+            result = parameterString != null;
+                // %1 := string.operator !=(string, string) [ string parameterString const ]
+                // result := __id [ %1 ]
+            Class1 classResult = some + other;
+                // %2 := Namespace.Class1.operator +(Namespace.Class1, Namespace.Class1) [ Namespace.Class1 some other ]
+                // classResult := __id [ %2 ]
+            classResult = classResult + some + other;
+                // %3 := Namespace.Class1.operator +(Namespace.Class1, Namespace.Class1) [ Namespace.Class1 classResult some ]
+                // %4 := Namespace.Class1.operator +(Namespace.Class1, Namespace.Class1) [ Namespace.Class1 %3 other ]
+                // classResult := __id [ %4 ]
+        }
+        public static Class1 operator+ (Class1 left, Class1 right)
+        {
+            return null;
         }
     }
 }";
