+++ sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/ControlFlowGraph/UcfgBuilder_Instructions_Suspect.cs
@@ -78,5 +78,32 @@ public class Foo
 }";
             UcfgVerifier.VerifyInstructions(code, "Bar");
         }
+
+        [TestMethod]
+        public void Dynamic()
+        {
+            const string code = @"
+namespace Namespace
+{
+    public class Class1
+    {
+        public string Foo(Class1 c)
+        {
+            dynamic dyn = c.Bar;
+                // dyn := __id [ const ]
+
+            if (dyn.User != null)
+                // %0 := dynamic.operator !=(dynamic, dynamic) [ __unknown const const ]
+            {
+                return ""bar"";
+            }
+
+            return c.ToString();
+                // %1 := object.ToString() [ c ]
+        }
+    }
+}";
+            UcfgVerifier.VerifyInstructions(code, "Foo");
+        }
     }
 }
